Add GetMetadata API; implement for keychain+file
A new `GetMetadata()` API exports a new `Metadata` type.  The `Metadata`
items can embed a `*Item`, which is permitted to be nil if no `Item`
fields at all can be retrieved without prompting.

`GetMetadata` should always be safe to call without needing to
authenticate and thus is suitable for use in batch workflows.

Implement `GetMetadata` for file and keychain backends.

Note: the keychain backend requires an updated `go-keychain` library,
not yet vendored into this repo, and requires Go 1.10 to build.