Add GetMetadata API
A new `GetMetadata()` API exports a new `Metadata` type.  The `Metadata`
items can embed a `*Item`, which is permitted to be nil if no `Item`
fields at all can be retrieved without prompting.

`GetMetadata` should always be safe to call without needing to
authenticate and thus is suitable for use in batch workflows.

Implement `GetMetadata` for file and keychain backends.
For Linux, stub out the call and return an error.

The code builds on Linux but GetMetadata is unusable; the libsecret API
potentially could be used, if coupled with a SetMetadata API too.

For keychain, this version now adapts to the variation of my patch which
keybase/go-keychain merged, where they switched the struct field name.