Fix removal of keys with the pass backend and add corresponding test
Prior to this commit, `aws-vault remove somekey` appeared to succeed but does not actually
remove anything from the pass password store. This is due to the cmd returned from the call to
the pass function never actually being run. Additionally, to avoid pass prompting the user
whether they want to remove the key, run with the `-f|--force` option. This implies that the
caller will take the appropriate steps and prompt/warn the user prior to key removal.

Signed-off-by: Collin J. Doering <collin@rekahsoft.ca>