add tests for libsecret (secret-service)
This commit adds test cases covering the expected behavior of the
libsecret backend based on the macos keychain implementation as the
reference implementation. In particular this means the tests do not
currently pass. The next commit will include API changes in libsecret.go
to align the behavior to keychain backend, such as returning
`ErrKeyNotFound` instead of a general libsecret error about a missing
collection when `Get()`.

NOTE: These tests are not runnable from a headless environment such as
Docker or a CI pipeline due to the DBus "prompt" interface being called
when creating and unlocking a keychain.