add test coverage to macos keychain interface
Adds coverage for Get/List/Remove when the keychain does not yet exist
as well as when it does already exist, such as having called .Set() at
least once.

This is useful to help identify API behavior changes within the keychain
backend as well as helping ensure consistent behavior across the other
backends.