Add in memory keyring.
An in-memory only keyring has been added. This will be useful for us because
we have command line tools that we want to store longer term state in
something like the OSX keychain, but later we may wish to automate the utility
in a more transient, ephemeral way. In that case, the memory store will store
the data for the duration of the run and recreate them on subsequent runs.

I've additionally refactored slightly and added in a test for an Item.Equals
function.