go.mod: use forked go-keychain
This updates the go modules (go.mod) to directly use the forked 99designs/go-keychain repository,
instead of requiring keybase/go-keychain and then using the replace directive.  The reasoning for
this is because the replace directive does not propagate to projects which depend on the go-keychain
project.  This means other projects which depend on the go-keychain project still need to add their
own replace directive.  Removing the replace directive from this project means that consumers
will automatically get the 99designs/go-keychain indirect dependency.  If there is a conflict
with another indirect dependency on keybase/go-keychain then that project can decide whether
to replace one or the other dependency.